basic-front-end-concepts-master/learning-app/requirements.txt
Django>=2.2
freezegun>=0.3.9
python-dateutil>=2.6.1
pytz>=2017.3
six>=1.11.0
djangorestframework>=3.9.2
Markdown>=3.0.1
django-filter>=2.0.0
drf-yasg>=1.14.0
python-keycloak>=0.23.0
model-mommy>=1.6.0
mock>=2.0.0
django-cors-headers>=2.4.0
pydub>=0.23.1
PyJWT>=2.0.1
